@@ -13,16 +13,16 @@ if (keystorePropertiesFile.exists() && keystorePropertiesFile.canRead()) {
 }
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "io.github.gmathi.novellibrary"
-        minSdkVersion 16
-        targetSdkVersion 28
+        minSdkVersion 21
+        targetSdkVersion 29
         multiDexEnabled true
-        versionCode 69
-        versionName "0.9.9.3.beta"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        versionCode 73
+        versionName "0.9.9.5.beta#patch1"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
     }
 
@@ -50,24 +50,23 @@ android {
     }
 }
 
-def support_lib_version = "28.0.0"
-
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
-    implementation "com.android.support:support-v4:$support_lib_version"
-    implementation "com.android.support:appcompat-v7:$support_lib_version"
-    implementation "com.android.support:cardview-v7:$support_lib_version"
-    implementation "com.android.support:design:$support_lib_version"
-    implementation "com.android.support:support-vector-drawable:$support_lib_version"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
+    implementation 'androidx.preference:preference-ktx:1.1.0'
 
-    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'androidx.multidex:multidex:2.0.1'
 
-    implementation 'org.jsoup:jsoup:1.11.2'
+    implementation 'org.jsoup:jsoup:1.12.1'
     implementation 'org.cryse.widget:persistentsearchview:1.0.4@aar'
     implementation 'org.greenrobot:eventbus:3.0.0'
 
@@ -106,13 +105,14 @@ dependencies {
         exclude group: 'com.android.support'
     })
 
-    implementation "com.google.android.gms:play-services-gcm:16.1.0"
-    implementation "com.google.android.gms:play-services-drive:16.1.0"
-    implementation "com.google.android.gms:play-services-auth:16.0.1"
-    implementation "com.google.firebase:firebase-messaging:17.6.0"
-    implementation "com.google.firebase:firebase-auth:16.2.1"
-    implementation "com.google.firebase:firebase-core:16.0.8"
-    implementation 'com.google.firebase:firebase-config:16.5.0'
+    implementation "com.google.android.gms:play-services-gcm:17.0.0"
+    implementation "com.google.android.gms:play-services-drive:17.0.0"
+    implementation "com.google.android.gms:play-services-auth:17.0.0"
+
+    implementation "com.google.firebase:firebase-messaging:20.1.0"
+    implementation "com.google.firebase:firebase-auth:19.2.0"
+    implementation "com.google.firebase:firebase-core:17.2.1"
+    implementation 'com.google.firebase:firebase-config:19.1.0'
     implementation 'com.google.code.gson:gson:2.8.5'
 
 
