@@ -1,11 +1,12 @@
 package io.github.gmathi.novellibrary.adapter
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 import io.github.gmathi.novellibrary.R
 import io.github.gmathi.novellibrary.util.inflate
 import java.util.*
+import java.util.concurrent.atomic.AtomicBoolean
 import kotlin.collections.ArrayList
 
 class GenericAdapter<T>(val items: ArrayList<T>, val layoutResId: Int, val listener: Listener<T>, var loadMoreListener: LoadMoreListener? = null) : RecyclerView.Adapter<GenericAdapter.ViewHolder<T>>() {
@@ -19,7 +20,12 @@ class GenericAdapter<T>(val items: ArrayList<T>, val layoutResId: Int, val liste
 
     override fun onBindViewHolder(holder: ViewHolder<T>, position: Int) = if (position == items.size) holder.loadMore(loadMoreListener) else holder.bind(item = items[position], listener = listener, position = position)
 
-    override fun onBindViewHolder(holder: ViewHolder<T>, position: Int, payloads: MutableList<Any>) = if (position == items.size) holder.loadMore(loadMoreListener) else holder.bind(item = items[position], listener = listener, position = position, payloads = payloads)
+    override fun onBindViewHolder(holder: ViewHolder<T>, position: Int, payloads: MutableList<Any>) {
+        if (items.size <= position + (loadMoreListener?.preloadCount ?: 0))
+            holder.loadMore(loadMoreListener)
+        if (position != items.size)
+            holder.bind(item = items[position], listener = listener, position = position, payloads = payloads)
+    }
 
     override fun getItemCount() = if (loadMoreListener != null) items.size + 1 else items.size
 
@@ -56,6 +62,8 @@ class GenericAdapter<T>(val items: ArrayList<T>, val layoutResId: Int, val liste
 
     interface LoadMoreListener {
         var currentPageNumber: Int
+        val preloadCount: Int
+        val isPageLoading: AtomicBoolean
         fun loadMore()
     }
 
