@@ -1,18 +1,19 @@
 package io.github.gmathi.novellibrary.activity.settings
 
 import android.annotation.SuppressLint
-import android.graphics.Rect
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.DividerItemDecoration
-import android.support.v7.widget.RecyclerView
 import android.view.MenuItem
 import android.view.View
+import android.widget.Toast
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.DividerItemDecoration
 import io.github.gmathi.novellibrary.R
 import io.github.gmathi.novellibrary.activity.BaseActivity
 import io.github.gmathi.novellibrary.adapter.GenericAdapter
-import io.github.gmathi.novellibrary.dataCenter
+import io.github.gmathi.novellibrary.util.Constants.SYSTEM_DEFAULT
 import io.github.gmathi.novellibrary.util.CustomDividerItemDecoration
+import io.github.gmathi.novellibrary.util.LocaleManager.Companion.changeLocale
+import io.github.gmathi.novellibrary.util.LocaleManager.Companion.translated
 import io.github.gmathi.novellibrary.util.applyFont
 import io.github.gmathi.novellibrary.util.setDefaults
 import kotlinx.android.synthetic.main.activity_language.*
@@ -22,8 +23,33 @@ import java.util.*
 
 class LanguageActivity : BaseActivity(), GenericAdapter.Listener<String> {
 
-    private val languagesMap = HashMap<String, String>()
-    private val languagesImageResourceMap = HashMap<String, Int>()
+    companion object {
+        private val languagesMap = HashMap<String, String>()
+        private val languagesImageResourceMap = HashMap<String, Int>()
+
+        init {
+            languagesMap["English"] = "en_US"
+            languagesMap["German"] = "de_DE"
+            languagesMap["Indonesian"] = "id_ID"
+            languagesMap["Portuguese"] = "pt_BR"
+            languagesMap["Spanish"] = "es_CL"
+            languagesMap["Turkish"] = "tr_TR"
+            languagesMap["French"] = "fr_FR"
+            languagesMap["Tagalog"] = "tr_PH"
+
+            languagesImageResourceMap["System Default"] = android.R.color.transparent
+            languagesImageResourceMap["English"] = R.drawable.flag_us
+            languagesImageResourceMap["German"] = R.drawable.flag_de
+            languagesImageResourceMap["Indonesian"] = R.drawable.flag_id
+            languagesImageResourceMap["Portuguese"] = R.drawable.flag_br
+            languagesImageResourceMap["Spanish"] = R.drawable.flag_cl
+            languagesImageResourceMap["Turkish"] = R.drawable.flag_tr
+            languagesImageResourceMap["French"] = R.drawable.flag_fr
+            languagesImageResourceMap["Tagalog"] = R.drawable.flag_tl
+        }
+    }
+
+    private var changeLanguage: Boolean = false
 
     lateinit var adapter: GenericAdapter<String>
 
@@ -31,34 +57,25 @@ class LanguageActivity : BaseActivity(), GenericAdapter.Listener<String> {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_language)
         setSupportActionBar(toolbar)
-        setLanguages()
+
+        if (intent.hasExtra("changeLanguage")) {
+            changeLanguage = intent.getBooleanExtra("changeLanguage", false)
+        }
+
+        if (savedInstanceState != null && savedInstanceState.containsKey("changeLanguage")) {
+            changeLanguage = savedInstanceState.getBoolean("changeLanguage")
+        }
+
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
         supportActionBar?.title = resources.getString(R.string.languages_supported)
         setRecyclerView()
     }
 
-    private fun setLanguages() {
-        languagesMap["English"] = "en_US"
-        languagesMap["German"] = "de_DE"
-        languagesMap["Indonesian"] = "id_ID"
-        languagesMap["Portuguese"] = "pt_BR"
-        languagesMap["Spanish"] = "es_CL"
-        languagesMap["Turkish"] = "tr_TR"
-        languagesMap["French"] = "fr_FR"
-        languagesMap["Tagalog"] = "tr_PH"
-
-        languagesImageResourceMap["English"] = R.drawable.flag_us
-        languagesImageResourceMap["German"] = R.drawable.flag_de
-        languagesImageResourceMap["Indonesian"] = R.drawable.flag_id
-        languagesImageResourceMap["Portuguese"] = R.drawable.flag_br
-        languagesImageResourceMap["Spanish"] = R.drawable.flag_cl
-        languagesImageResourceMap["Turkish"] = R.drawable.flag_tr
-        languagesImageResourceMap["French"] = R.drawable.flag_fr
-        languagesImageResourceMap["Tagalog"] = R.drawable.flag_tl
-    }
-
     private fun setRecyclerView() {
-        adapter = GenericAdapter(items = ArrayList(languagesMap.keys.sorted()), layoutResId = R.layout.listitem_image_title_subtitle, listener = this)
+        val items = ArrayList(languagesMap.keys.sorted())
+        if (changeLanguage)
+            items.add(0, resources.getString(R.string.system_default))
+        adapter = GenericAdapter(items = items, layoutResId = R.layout.listitem_image_title_subtitle, listener = this)
         recyclerView.setDefaults(adapter)
         recyclerView.addItemDecoration(CustomDividerItemDecoration(this, DividerItemDecoration.VERTICAL))
         swipeRefreshLayout.isEnabled = false
@@ -75,7 +92,15 @@ class LanguageActivity : BaseActivity(), GenericAdapter.Listener<String> {
 
     @SuppressLint("NewApi")
     override fun onItemClick(item: String) {
-        dataCenter.language = languagesMap[item]!!
+        val language = if (item == resources.getString(R.string.system_default)) SYSTEM_DEFAULT else languagesMap[item]!!.split("_")[0]
+        if (changeLanguage)
+            changeLocale(this, language)
+        else if (language != "en") {
+            val translated = translated(this, language)
+            val total = translated(this)
+            val percentage = String.format("%.2f", translated.toDouble() / total * 100)
+            Toast.makeText(this, resources.getString(R.string.translated, translated, total, percentage), Toast.LENGTH_LONG).show()
+        }
     }
 
     override fun onOptionsItemSelected(item: MenuItem?): Boolean {
@@ -83,4 +108,8 @@ class LanguageActivity : BaseActivity(), GenericAdapter.Listener<String> {
         return super.onOptionsItemSelected(item)
     }
 
+    override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        outState.putBoolean("changeLanguage", changeLanguage)
+    }
 }
