@@ -1,11 +1,11 @@
 package io.github.gmathi.novellibrary.fragment
 
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
 import co.metalab.asyncawait.async
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
@@ -22,11 +22,14 @@ import io.github.gmathi.novellibrary.util.setDefaults
 import kotlinx.android.synthetic.main.content_recycler_view.*
 import kotlinx.android.synthetic.main.listitem_novel.view.*
 import java.net.URLEncoder
+import java.util.concurrent.atomic.AtomicBoolean
 
 
 class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel>, GenericAdapter.LoadMoreListener {
 
     override var currentPageNumber: Int = 1
+    override val preloadCount:Int = 50
+    override val isPageLoading: AtomicBoolean = AtomicBoolean(false)
     private lateinit var searchTerm: String
     private lateinit var resultType: String
 
@@ -96,7 +99,7 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel>, Gener
             var results: ArrayList<Novel>? = null
 
             when (resultType) {
-                HostNames.NOVEL_UPDATES -> results = await { NovelApi.searchNovelUpdates(searchTerms, currentPageNumber) }
+                HostNames.NOVEL_UPDATES -> results = await { NovelApi.searchNovelUpdates_New(searchTerms, currentPageNumber) }
                 HostNames.ROYAL_ROAD -> results = await { NovelApi.searchRoyalRoad(searchTerms, currentPageNumber) }
                 HostNames.NOVEL_FULL -> results = await { NovelApi.searchNovelFull(searchTerms, currentPageNumber) }
                 HostNames.WLN_UPDATES -> results = await { NovelApi.searchWlnUpdates(searchTerms) }
@@ -106,6 +109,7 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel>, Gener
             if (results != null) {
                 if (isVisible && (!isDetached || !isRemoving)) {
                     loadSearchResults(results)
+                    isPageLoading.lazySet(false)
                     swipeRefreshLayout.isRefreshing = false
                 }
             } else {
@@ -145,8 +149,10 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel>, Gener
     }
 
     override fun loadMore() {
-        currentPageNumber++
-        searchNovels()
+        if (isPageLoading.compareAndSet(false, true)) {
+            currentPageNumber++
+            searchNovels()
+        }
     }
 
     //region Adapter Listener Methods - onItemClick(), viewBinder()
@@ -185,7 +191,7 @@ class SearchTermFragment : BaseFragment(), GenericAdapter.Listener<Novel>, Gener
 
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        if (adapter.items.isNotEmpty())
+        if (::adapter.isInitialized && adapter.items.isNotEmpty())
             outState.putSerializable("results", adapter.items)
     }
 
