@@ -8,7 +8,7 @@ fun NovelApi.searchUrl(url: String, pageNumber: Int): ArrayList<Novel>? {
         //val host = URI(url).host
         when {
             url.contains(HostNames.ROYAL_ROAD_OLD) || url.contains(HostNames.ROYAL_ROAD) -> return searchRoyalRoadUrl(url)
-            url.contains(HostNames.NOVEL_UPDATES) -> return searchNovelUpdatesUrl(url, pageNumber)
+            url.contains(HostNames.NOVEL_UPDATES) -> return searchNovelUpdatesUrl_New(url, pageNumber)
             url.contains(HostNames.WLN_UPDATES) -> return searchWlnUpdatesUrl(url)
             url.contains(HostNames.NOVEL_FULL) -> return searchNovelFullUrl(url)
         }
@@ -58,6 +58,31 @@ fun NovelApi.searchNovelUpdatesUrl(searchUrl: String, pageNumber: Int): ArrayLis
     }
     return searchResults
 }
+
+fun NovelApi.searchNovelUpdatesUrl_New(searchUrl: String, pageNumber: Int): ArrayList<Novel>? {
+    var searchResults: ArrayList<Novel>? = null
+    try {
+        searchResults = ArrayList()
+        val document = getDocumentWithUserAgent(searchUrl.addPageNumberToUrl(pageNumber, "pg"))
+        val elements = document.body().select("div.search_main_box_nu") ?: return searchResults
+        for (element in elements) {
+            val novelName = element.selectFirst("div.search_title > a")?.text() ?: continue
+            val novelUrl = element.selectFirst("div.search_title > a")?.attr("abs:href") ?: continue
+            val novel = Novel(novelName, novelUrl)
+            novel.imageUrl = element.selectFirst("div.search_img_nu > img[src]")?.attr("abs:src")
+
+//            val ratingsElement = element.selectFirst("div.search_ratings") ?: continue
+//            ratingsElement.children().remove()
+            novel.rating = element.select("span.search_ratings").text().trim().replace("(", "").replace(")", "")
+            searchResults.add(novel)
+        }
+
+    } catch (e: Exception) {
+        e.printStackTrace()
+    }
+    return searchResults
+}
+
 
 fun NovelApi.searchWlnUpdatesUrl(url: String): ArrayList<Novel>? {
     var searchResults: ArrayList<Novel>? = null
