@@ -8,8 +8,7 @@ object Constants {
     const val BACKUP_DIR = "NovelLibraryApp-Backup"
     const val SCROLL_LENGTH = 30
     const val FILE_PROTOCOL = "file://"
-
-
+    const val SYSTEM_DEFAULT = "systemDefault"
 
     //region DownloadQueue Status Values
 
